# src/pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The IPEHR Gateway is an openEHR compliant EHR server implementation that stores encrypted medical data in a Filecoin distributed file storage.",
        "title": "IPEHR Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://bsn.si/blockchain",
            "email": "support@bsn.si"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.2"
    },
    "host": "gateway.ipehr.org",
    "basePath": "/v1",
    "paths": {
        "/access/document": {
            "post": {
                "description": "Sets access to the document with the specified CID for the user with the userID.\nPossible access levels: `owner`, `admin`, `read`, `noAccess`",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ACCESS"
                ],
                "summary": "Set user access to the document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    },
                    {
                        "description": "DTO with data to create group access",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DocAccessSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Indicates that the request to change the level of access to the document was successfully created"
                    },
                    "400": {
                        "description": "Is returned when the request has invalid content."
                    },
                    "404": {
                        "description": "Is returned when the userID for which access is set is not found "
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/access/document/": {
            "get": {
                "description": "Returns the list of documents available to the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ACCESS"
                ],
                "summary": "Get a document access list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DocAccessListResponse"
                        }
                    },
                    "400": {
                        "description": "Is returned when the request has invalid content."
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DocAccessDocument": {
            "type": "object",
            "properties": {
                "CID": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                }
            }
        },
        "model.DocAccessDocumentGroup": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DocAccessDocument"
                    }
                },
                "groupID": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "parentGroupID": {
                    "type": "string"
                }
            }
        },
        "model.DocAccessListResponse": {
            "type": "object",
            "properties": {
                "documentGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DocAccessDocumentGroup"
                    }
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DocAccessDocument"
                    }
                }
            }
        },
        "model.DocAccessSetRequest": {
            "type": "object",
            "properties": {
                "accessLevel": {
                    "type": "string"
                },
                "cid": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}